+++ SourceCode/e_commerce_app/lib/screens/home/home_screen.dart
@@ -50,24 +50,4 @@ class HomeScreen extends StatelessWidget {
   }
 }
 
-class SectionTitle extends StatelessWidget {
-  final String title;
-  const SectionTitle({
-    Key? key,
-    required this.title,
-  }) : super(key: key);
 
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.symmetric(horizontal: 20.0),
-      child: Align(
-        alignment: Alignment.topLeft,
-        child: Text(
-          title,
-          style: Theme.of(context).textTheme.headline6,
-        ),
-      ),
-    );
-  }
-}
