+++ SourceCode/e_commerce_app/lib/widgets/widgets.dart
@@ -1,3 +1,6 @@
 export 'custom_appbar.dart';
 export 'custom_navbar.dart';
 export 'hero_carousel_card.dart';
+export 'package:e_commerce_app/widgets/section_title.dart';
+
+
